@@ -2,7 +2,9 @@ package br.com.prodplus.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -20,6 +22,7 @@ import br.com.prodplus.services.ConfiguracaoService;
  */
 @RestController
 @RequestMapping("/configuracoes")
+@CrossOrigin("http://localhost:4200")
 public class ConfiguracaoController {
 
 	@Autowired
@@ -36,7 +39,8 @@ public class ConfiguracaoController {
 	}
 
 	@PutMapping("/{periodo}")
-	public ResponseEntity<Configuracao> inserirTurno(String periodo, @RequestBody Turno turno) {
+	public ResponseEntity<Configuracao> inserirTurno(@PathVariable String periodo,
+			@RequestBody Turno turno) {
 		return ResponseEntity.ok(this.configService.inserirTurno(periodo, turno));
 	}
 
