@@ -1,7 +1,13 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
+import { ConfiguracaoComponent } from './configuracao/configuracao.component';
+import { HomeComponent } from './home/home.component';
 
-const routes: Routes = [];
+const routes: Routes = [
+  { path: '', redirectTo: 'home', pathMatch: 'full' },
+  { path: 'home', component: HomeComponent },
+  { path: 'configuracoes', component: ConfiguracaoComponent }
+];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
